requirements.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
google-api-core==2.11.0
google-auth==2.16.2
google-cloud==0.34.0
google-cloud-secret-manager==2.16.0
google-crc32c==1.5.0
googleapis-common-protos==1.58.0
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.51.3
grpcio-status==1.51.3
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==1.0.2
requests==2.28.2
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.6
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
